# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HashWatch API",
    "description": "Backend API for HashWatch \u2014 monitor and manage cryptocurrency miners in real-time.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://{host}:{port}/api",
      "variables": {
        "host": {
          "default": "localhost",
          "description": "API host"
        },
        "port": {
          "default": "4000",
          "description": "API port"
        }
      }
    }
  ],
  "paths": {
...
    "/proxy": {
      "post": {
        "tags": ["Proxy"],
        "summary": "Proxy an HTTP request to a miner",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProxyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Miner response data",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "URL is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProxyErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream error or miner unreachable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProxyErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/proxy/restart": {
      "post": {
        "tags": ["Proxy"],
        "summary": "Send restart command to a miner",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProxyRestartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Restart command sent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "502": {
            "description": "Could not reach miner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProxyErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/proxy/flash": {
      "post": {
        "tags": ["Proxy"],
        "summary": "Flash firmware to a miner (OTA update)",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": { "type": "string", "description": "Miner OTA endpoint URL" },
                  "method": { "type": "string", "default": "POST" },
                  "body": { "type": "string", "description": "JSON body with firmware URL" }
                },
                "required": ["url"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Firmware flash initiated" },
          "400": { "description": "Missing url" },
          "403": { "description": "Forbidden URL" },
          "502": { "description": "Miner unreachable or flash failed" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /auth/register or /auth/login. Expires in 30 days."
      }
    },
    "schemas": {
...
      "ProxyRequest": {
        "type": "object",
        "required": ["url"],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Target miner URL"
          },
          "method": {
            "type": "string",
            "default": "GET",
            "enum": ["GET", "POST", "PUT", "DELETE", "PATCH"]
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Headers to forward (blocked: host, authorization, cookie, etc.)"
          },
          "data": {
            "description": "Request body to forward"
          }
        }
      },
      "ProxyRestartRequest": {
        "type": "object",
        "required": ["url"],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Target miner URL"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "ProxyErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "upstreamStatus": {
            "type": "integer"
          }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [true]
          }
        }
      },
...
    }
  },
...
}